# services/procurement-service/openAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Procurement Service APIs",
    "version": "1.0.0",
    "description": "API documentation for Procurement Service. This service manages purchase orders, supplier catalogs mirror, and procurement workflows."
  },
  "servers": [
    {
      "url": "http://localhost:8002"
    }
  ],
  "components": {
    "schemas": {
      "POAudit": {
        "poNumber": {
          "example": "PO-2026-0001"
        },
        "fromStatus": {
          "type": "string",
          "enum": ["DRAFT", "SUBMITTED", "APPROVED", "REJECTED"],
          "example": "DRAFT"
        },
        "toStatus": {
          "type": "string",
          "enum": ["DRAFT", "SUBMITTED", "APPROVED", "REJECTED"],
          "example": "DRAFT"
        },
        "action": {
          "type": "string",
          "example": "STATUS_UPDATE_SUBMITTED"
        }
      },
      "PurchaseOrder": {
        "type": "object",
        "required": ["poNumber", "requester", "neededBy"],
        "properties": {
          "_id": {
            "type": "string"
          },
          "poNumber": {
            "type": "string",
            "example": "PO-2026-0001"
          },
          "supplier": {
            "type": "string",
            "example": "Flexitallic"
          },
          "requester": {
            "type": "string",
            "example": "Buyer Name"
          },
          "costCenter": {
            "type": "string",
            "example": "CC-1234"
          },
          "neededBy": {
            "type": "string",
            "format": "date",
            "example": "2026-03-01"
          },
          "paymentTerms": {
            "type": "string",
            "enum": ["Cash", "Credit", "Installments"],
            "example": "Credit"
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "SUBMITTED", "APPROVED", "REJECTED"],
            "example": "DRAFT"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineItem"
            }
          },
          "totalAmount": {
            "type": "number",
            "example": 227.5
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PurchaseOrderItem": {
        "data": {
          "type": "object",
          "items": {
            "$ref": "#/components/schemas/PurchaseOrder"
          }
        },
        "history": {
          "type": "object",
          "items": {
            "$ref": "#/components/schemas/POAudit"
          }
        }
      },
      "CreateDraftPO": {
        "type": "object",
        "required": ["requester", "neededBy"],
        "properties": {
          "requester": {
            "type": "string",
            "example": "Buyer Name"
          },
          "costCenter": {
            "type": "string",
            "example": "CC-1234"
          },
          "neededBy": {
            "type": "string",
            "format": "date",
            "example": "2026-03-01"
          },
          "paymentTerms": {
            "type": "string",
            "enum": ["Cash", "Credit", "Installments"]
          }
        }
      },
      "LineItemRequest": {
        "type": "object",
        "required": ["catalogItemId", "quantity"],
        "properties": {
          "catalogItemId": {
            "type": "string",
            "example": "GST-0001"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "example": 5
          }
        }
      },
      "LineItem": {
        "type": "object",
        "required": ["catalogItemId", "quantity"],
        "properties": {
          "catalogItemId": {
            "type": "string",
            "example": "GST-0001"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "example": 5
          },
          "name": {
            "type": "string",
            "example": "Spiral Wound Gasket"
          },
          "pricePerUnit": {
            "type": "number",
            "example": 45.5
          },
          "priceTotal": {
            "type": "number",
            "example": 227.5
          }
        }
      },
      "UpdateStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["SUBMITTED", "APPROVED", "REJECTED"]
          },
          "notes": {
            "type": "string",
            "example": "Budget approved by finance"
          }
        }
      },
      "CatalogMirror": {
        "type": "object",
        "required": ["productId", "name", "supplier", "priceUsd", "inStock"],
        "properties": {
          "_id": {
            "type": "string"
          },
          "productId": {
            "type": "string",
            "example": "GST-0001"
          },
          "name": {
            "type": "string",
            "example": "Spiral Wound Gasket"
          },
          "supplier": {
            "type": "string",
            "example": "Flexitallic"
          },
          "priceUsd": {
            "type": "number",
            "example": 45.5
          },
          "inStock": {
            "type": "boolean",
            "example": true
          },
          "leadTimeDays": {
            "type": "number",
            "example": 5
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SyncCatalogMirrorRequest": {
        "type": "object",
        "required": [
          "productId",
          "name",
          "supplier",
          "priceUsd",
          "inStock",
          "leadTimeDays"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "example": "GST-0001"
          },
          "name": {
            "type": "string",
            "example": "Spiral Wound Gasket"
          },
          "supplier": {
            "type": "string",
            "example": "Flexitallic"
          },
          "priceUsd": {
            "type": "number",
            "example": 45.5
          },
          "inStock": {
            "type": "boolean",
            "example": true
          },
          "leadTimeDays": {
            "type": "number",
            "example": 5
          }
        }
      }
    }
  },
  "paths": {
    "/api/v1/mirror-catalog/": {
      "get": {
        "summary": "Get mirrored catalog items",
        "description": "Returns locally cached catalog items synchronized from catalog service. This endpoint is optimized for fast read access inside the procurement system and avoids cross-service latency.",
        "tags": ["Catalog Mirror"],
        "responses": {
          "200": {
            "description": "List of mirrored items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogMirror"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Sync catalog item into mirror",
        "description": "Creates or updates a mirrored catalog item (used by catalog service or message consumer). This endpoint behaves as an UPSERT operation.",
        "tags": ["Catalog Mirror"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncCatalogMirrorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing item updated successfully"
          },
          "201": {
            "description": "New item created in mirror"
          },
          "400": {
            "description": "Invalid request body"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/mirror-catalog/{id}": {
      "delete": {
        "summary": "Remove mirrored catalog item",
        "description": "Deletes item from local mirror when removed from catalog service to maintain eventual consistency.",
        "tags": ["Catalog Mirror"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "GST-0001"
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed from mirror"
          },
          "404": {
            "description": "Item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/purchase-order/": {
      "post": {
        "summary": "To create a purchase order with status Draft.",
        "description": "This is the first step to create a purchase order. It assigns a PO number and sets the status to DRAFT. The PO can then be modified by adding items before submission.",
        "tags": ["Purchase Orders"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftPO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft PO created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrder"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get purchase orders",
        "description": "Supports filtering by status, supplier, requester and sorting by various fields. This allows users to easily find and manage their purchase orders based on workflow state and other criteria.",
        "tags": ["Purchase Orders"],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "SUBMITTED", "APPROVED", "REJECTED"]
            }
          },
          {
            "in": "query",
            "name": "supplier",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "requester",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "poNumber",
                "supplier",
                "status",
                "totalAmount",
                "createdAt"
              ]
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of purchase orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurchaseOrder"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/purchase-order/{id}": {
      "get": {
        "summary": "Get purchase order by ID along with history",
        "description": "Retrieves a specific purchase order by its ID, including all details and a history of status changes. This allows users to track the lifecycle of the PO and understand its current state in the procurement process.",
        "tags": ["Purchase Orders"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Purchase order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrderItem"
                }
              }
            }
          },
          "404": {
            "description": "PO not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/purchase-order/{id}/items": {
      "post": {
        "summary": "Add item to purchase order",
        "description": "Adds an item to a draft PO. Returns 409 if supplier conflict occurs.",
        "tags": ["Purchase Orders"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrder"
                }
              }
            }
          },
          "400": {
            "description": "Invalid state (cannot modify non-draft PO)"
          },
          "404": {
            "description": "Purchase order not found"
          },
          "409": {
            "description": "Supplier conflict detected"
          },
          "422": {
            "description": "Invalid quantity or pricing"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/purchase-order/{id}/status": {
      "patch": {
        "summary": "Update purchase order status",
        "description": "Changes PO workflow state (DRAFT → SUBMITTED → APPROVED/REJECTED)",
        "tags": ["Purchase Orders"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrder"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition"
          },
          "404": {
            "description": "Purchase order not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "tags": []
}